+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 ### Изменено
 
 - выделена задача для установки nuget
+- добавлена задача сборки 'distclean' в целях соответствия
+  [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)
 - переименованы задачи сборки в целях соответствия
   [стандартным целям make](https://www.gnu.org/software/make/manual/html_node/Standard-Targets.html)